# [TIR][Hexagon] Enhancement of NarrowDataType pass for binary ops